+++ ChangeLog.md
@@ -217,7 +217,7 @@
 - Buttons können jetzt zu Reactbuttons gemacht werden
   - Die Buttons in den Modals wurden entsprechend angepasst
 
-#### v2.5.11 - GitHub Actions, Test and UI Improvements
+#### v2.5.11 - GitHub Actions und Keyboard Control Improvements
 - GitHub Actions um Tests zu Automatisieren
 - WeekTime Modal angepasst
   - mehr Breite weniger Höhe, bessere Verteilung der Komponenten
@@ -226,3 +226,26 @@
   - Type Benennung angepasst
   - Funktion in andere Datei verschoben
 - bugfix: restliche FlexOffice Zeit diesen Monat
+- bugfix: Die DevOptions können jetzt wieder über eine bestimmte URL automatisch gestartet werden
+- type float entfernt (css wird ist jetzt an die id gebunden)
+- besserer Text bei der CurrentStatsMessageWithValues
+- bessere Keyboard Control (Beschreibung entsprechen angepasst)
+  - F1 und F2 getauscht
+  - F3: +14 Min Gleitzeit
+  - F4: OverTimeAutomatic on/off
+  - Shift + G: zum Gleitzeitfeld fokussieren
+  - Shift + S: Six-Hour-Mode aktivieren
+  - Alt + w: WeekTimeCalculator öffnen
+  - Alt + f: FlexOfficeCalculator öffnen
+  - Alt + c: CurrentStatsMessageWithValues öffnen
+  - Alt + d: DevOptionsForm öffnen
+  - F3, F4 und Alt + d funktionieren nur noch wenn die DevOptions aktiv sind
+- new Utility Function (addTimeValues())
+    - entsprechende Tests
+    - Test für subtractTimeValues() verbessert
+- Zeit hinzufügen Button bei FlexOffice eingebaut, damit man einfach Zeiten ergänzen kann
+- getNumberFromFloat() gefixt, falls "" als Wert gefunden wird
+- Man kann jetzt FlexOffice Zeit auf einem extra Modal hinzufügen
+- bessere und mehr Testcases
+- Help Page eingebaut (alle Shortcuts werden aufgelistet)
+    - In einer Tabelle mit Shortcut, Action und Condition